app/src/test/java/com/makentoshe/habrachan/common/model/network/posts/GetPostsConverterFactoryTest.kt
package com.makentoshe.habrachan.common.model.network.posts

import com.makentoshe.habrachan.common.model.network.Result
import com.makentoshe.habrachan.resources.testResourcesDirectory
import io.mockk.every
import io.mockk.mockk
import okhttp3.ResponseBody
import org.junit.Assert.*
import org.junit.Before
import org.junit.Test
import retrofit2.Converter
import java.io.File

class GetPostsConverterFactoryTest {

    private val factory = GetPostsConverterFactory()
    private lateinit var converter: Converter<ResponseBody, Result.GetPostsResponse>

    @Before
    fun init() {
        converter = factory.responseBodyConverter(Result.GetPostsResponse::class.java, arrayOf(), mockk())!!
    }

    @Test
    fun `converter should be null for incompatible type`() {
        assertNull(factory.responseBodyConverter(Any::class.java, arrayOf(), mockk()))
    }

    @Test
    fun `should parse GetPostsSuccess`() {
        val successJson = File(testResourcesDirectory, "GetPostsSuccess.json")

        val responseBody = mockk<ResponseBody>()
        every { responseBody.string() } returns successJson.readText()

        converter.convert(responseBody).also {
            assertNotNull(it.success)
            assertNull(it.error)
        }
    }

    @Test
    fun `should parse ErrorResult`() {
        val errorJson = File(testResourcesDirectory, "ErrorResult.json")

        val responseBody = mockk<ResponseBody>()
        every { responseBody.string() } returns errorJson.readText()

        converter.convert(responseBody).also {
            assertNotNull(it.error)
            assertNull(it.success)
        }
    }

    @Test
    fun `should parse GetPostsError`() {
        val errorJson = File(testResourcesDirectory, "GetPostsError.json")

        val responseBody = mockk<ResponseBody>()
        every { responseBody.string() } returns errorJson.readText()

        converter.convert(responseBody).also {
            assertNotNull(it.error)
            assertNull(it.success)
        }
    }
}

